More doc fixes
