+++ README.md
@@ -17,6 +17,8 @@ Ramify is a Rust library for generating *branch diagrams* to visualize hierarchi
 ```
 See the [gallery](https://github.com/alexrutar/ramify#gallery) for more images.
 
+To use this library, you simply describe your graph structure and associated metadata, and the branch diagram is generated automatically in an optimal way.
+
 This library is specifically designed for ordered data: this library generates output similar to `git log --graph --all`, rather than the output of `tree`.
 A prototypical application is to visualize the undo-tree of a text file.
 The order is the timestamp of the edit, and the tree structure results from the undo relation.
@@ -34,7 +36,7 @@ The order is the timestamp of the edit, and the tree structure results from the 
 Interested?
 Check out the [API documentation](https://docs.rs/ramify/latest/ramify/) for more detail!
 
-Basic examples can be found in the [examples folder](https://github.com/alexrutar/ramify/tree/master/examples).
+Usage examples can be found in the [examples folder](https://github.com/alexrutar/ramify/tree/master/examples).
 
 The output generated by this crate is [highly configurable](https://docs.rs/ramify/latest/ramify/writer/).
 Explore the various styles and configuration options with the `gallery` example:
