+++ src/writer.rs
@@ -9,12 +9,17 @@
 //! [`FmtWriter`] which holds a [`fmt::Write`] implementation.
 //! The appearance of the diagram generated by these writers is specified by the [`Style`] struct.
 //!
+//! Note that layout configuration is found in the [`Config`](crate::Config) struct. Essentially,
+//! configuration which results in a different sequence of calls to the [`DiagramWrite`]
+//! methods belongs in [`Config`](crate::Config).
+//!
+//! ### Composing diagram writers
+//!
 //! If you want to customize the diagram writer, in many simple cases it is also possible to wrap an
-//! internal diagram writer.
+//! internal diagram writer. The [`DiagramWrite::write_fmt`] helps to compose writers, and also
+//! can make it simpler to write code which is generic over an arbitrary diagram writer.
 //!
-//! ### Custom diagram writer examples
-//!
-//! In the following example, we modify an [`IOWriter`] to print line numbers before each annotation
+//! In the following example, we wrap a diagram writer to print line numbers before each annotation
 //! line.
 //! ```
 //! use ramify::writer::{Branch, DiagramWrite, IOWriter, MergeBranch};
@@ -69,8 +74,10 @@
 //! ```
 mod branch;
 mod shim;
+mod style;
 
 pub use self::branch::{Branch, MergeBranch};
+pub use self::style::{Charset, Style};
 use shim::{Shim, WriteInner};
 
 use std::{fmt, io};
@@ -173,39 +180,27 @@ pub trait DiagramWrite {
         required: usize,
     ) -> Result<(), Self::Error>;
 
-    /// Write the text of the annotation line.
+    /// Write the text of a single annotation line.
     ///
     /// The index `idx` is the index of the annotation line in the block of annotations.
     fn write_annotation(&mut self, idx: usize, line: &str) -> Result<(), Self::Error>;
 
     /// Write a newline.
     ///
-    /// This is called to terminate branch diagram lines which do not contain annotations.
+    /// This is called when the diagram row is completed.
     fn write_newline(&mut self) -> Result<(), Self::Error>;
-
-    /// Prepare and then write an annotation followed by a newline.
-    ///
-    /// The default implementation calls, in order:
-    ///
-    /// 1. [`prepare_annotation`](DiagramWrite::prepare_annotation),
-    /// 2. [`write_annotation`](DiagramWrite::write_annotation)
-    /// 3. [`write_newline`](DiagramWrite::write_newline).
-    fn write_annotation_line(
-        &mut self,
-        idx: usize,
-        written: usize,
-        required: usize,
-        line: &str,
-    ) -> Result<(), Self::Error> {
-        self.prepare_annotation(idx, written, required)?;
-        self.write_annotation(idx, line)?;
-        self.write_newline()
-    }
 }
 
 /// A diagram writer which writes into an [`io::Write`] implementation.
 ///
-/// Also see the [`Style::io_writer`] method.
+/// Note that this writer is *not buffered*, and a diagram writer emits a very large number of
+/// small writes to the internal writer. It is strongly recommended to provide a buffered writer,
+/// for example with a [`BufWriter`](io::BufWriter) or a [`LineWriter`](io::LineWriter). Some
+/// built-in writers, such as [standard output](io::stdout), are automatically line buffered.
+///
+/// The [`DiagramWrite`] implementation will never flush output. You can flush the internal writer
+/// with [`flush`](Self::flush). Since a newline is always written after every diagram row, it is
+/// not necessary to flush if you are using a line-buffered writer.
 pub struct IOWriter<W> {
     /// The style.
     style: Style,
@@ -216,6 +211,9 @@ pub struct IOWriter<W> {
 impl<W> IOWriter<W> {
     /// Construct a new diagram writer with the provided style and internal [`io::Write`]
     /// implementation.
+    ///
+    /// Also see the [`Style::io_writer`] method.
+    #[inline]
     pub const fn new(style: Style, writer: W) -> Self
     where
         W: io::Write,
@@ -223,7 +221,18 @@ impl<W> IOWriter<W> {
         Self { style, writer }
     }
 
+    /// Flush the inner writer.
+    ///
+    /// This is identical to `self.writer().flush()`.
+    pub fn flush(&mut self) -> io::Result<()>
+    where
+        W: io::Write,
+    {
+        self.writer.flush()
+    }
+
     /// Get a mutable reference to the internal writer.
+    #[inline]
     pub const fn writer(&mut self) -> &mut W {
         &mut self.writer
     }
@@ -232,11 +241,13 @@ impl<W> IOWriter<W> {
     ///
     /// If you must modify the style, first decompose into [components](Self::components) and then rebuild
     /// this struct.
+    #[inline]
     pub const fn style(&self) -> &Style {
         &self.style
     }
 
     /// Recover the style and the internal writer.
+    #[inline]
     pub fn components(self) -> (Style, W) {
         (self.style, self.writer)
     }
@@ -337,6 +348,8 @@ impl<W: io::Write> DiagramWrite for IOWriter<W> {
 }
 
 /// A diagram writer which writes into a [`fmt::Write`] implementation.
+///
+/// Often, this is a wrapper around a string buffer.
 pub struct FmtWriter<W> {
     /// The style.
     style: Style,
@@ -349,6 +362,7 @@ impl<W> FmtWriter<W> {
     /// implementation.
     ///
     /// Also see the [`Style::fmt_writer`] method.
+    #[inline]
     pub const fn new(style: Style, writer: W) -> Self
     where
         W: fmt::Write,
@@ -357,6 +371,7 @@ impl<W> FmtWriter<W> {
     }
 
     /// Get a mutable reference to the internal writer.
+    #[inline]
     pub const fn writer(&mut self) -> &mut W {
         &mut self.writer
     }
@@ -365,11 +380,13 @@ impl<W> FmtWriter<W> {
     ///
     /// If you must modify the style, first decompose into [components](Self::components) and then rebuild
     /// this struct.
+    #[inline]
     pub const fn style(&self) -> &Style {
         &self.style
     }
 
     /// Recover the style and the internal writer.
+    #[inline]
     pub fn components(self) -> (Style, W) {
         (self.style, self.writer)
     }
@@ -458,392 +475,5 @@ impl<W: fmt::Write> DiagramWrite for FmtWriter<W> {
     #[inline]
     fn write_newline(&mut self) -> Result<(), Self::Error> {
         Shim(self).write_newline()
-    }
-}
-
-/// Configuration for the appearance of the branch diagram.
-///
-/// This struct is used by the [`IOWriter`] and [`FmtWriter`] implementations of [`DiagramWrite`].
-/// For more fine-grained configuration, one should manually implement [`DiagramWrite`].
-#[derive(Debug, Clone, PartialEq, Eq)]
-pub struct Style {
-    /// A set of characters to use in the diagram.
-    charset: Charset,
-    /// The margin at the beginning of each annotation.
-    annotation_margin: usize,
-    /// The minimum left alignment of each annotation line.
-    annotation_justification: usize,
-    /// The gap between vertices.
-    gutter_width: usize,
-    /// Put merge lines on top.
-    merge_over: bool,
-}
-
-impl Style {
-    /// Initialize a new style from the provided character set.
-    pub const fn new(charset: Charset) -> Self {
-        Self {
-            charset,
-            annotation_margin: 1,
-            annotation_justification: 0,
-            gutter_width: 0,
-            merge_over: false,
-        }
-    }
-
-    /// Get an [`IOWriter`] using this style.
-    pub const fn io_writer<W: io::Write>(self, writer: W) -> IOWriter<W> {
-        IOWriter::new(self, writer)
-    }
-
-    /// Get a [`FmtWriter`] using this style.
-    pub const fn fmt_writer<W: fmt::Write>(self, writer: W) -> FmtWriter<W> {
-        FmtWriter::new(self, writer)
-    }
-
-    /// Initialize with the [rounded corners](Charset::rounded_corners) character set.
-    ///
-    /// The charset looks as follows.
-    /// ```txt
-    /// 0    
-    /// ├┬╮  
-    /// │1├╮
-    /// ││2│
-    /// │3│├╮
-    /// │╭╯││
-    /// ││╭┤│
-    /// │││4│
-    /// ││5╭╯
-    /// │6╭╯
-    /// 7╭╯  
-    ///  8   
-    /// ```
-    pub const fn rounded_corners() -> Self {
-        Self::new(Charset::rounded_corners())
-    }
-
-    /// Initialize with the [sharp corners](Charset::sharp_corners) character set.
-    ///
-    /// The charset looks as follows.
-    /// ```txt
-    /// 0
-    /// ├┬┐
-    /// │1├┐
-    /// ││2│
-    /// │3│├┐
-    /// │┌┘││
-    /// ││┌┤│
-    /// │││4│
-    /// ││5┌┘
-    /// │6┌┘
-    /// 7┌┘
-    ///  8
-    /// ```
-    pub const fn sharp_corners() -> Self {
-        Self::new(Charset::sharp_corners())
-    }
-
-    /// Initialize with the [doubled lines](Charset::doubled_lines) character set.
-    ///
-    /// The charset looks as follows.
-    /// ```txt
-    /// 0
-    /// ╠╦╗
-    /// ║1╠╗
-    /// ║║2║
-    /// ║3║╠╗
-    /// ║╔╝║║
-    /// ║║╔╣║
-    /// ║║║4║
-    /// ║║5╔╝
-    /// ║6╔╝
-    /// 7╔╝
-    ///  8
-    /// ```
-    pub const fn doubled_lines() -> Self {
-        Self::new(Charset::doubled_lines())
-    }
-
-    /// Reset the character set used in the style.
-    pub const fn charset(mut self, charset: Charset) -> Self {
-        self.charset = charset;
-        self
-    }
-
-    /// Invert the internal character set.
-    ///
-    /// This inverts vertically by swapping characters like `╯` and `╮`. Calling this method twice
-    /// will return the original character set. This makes the lines connect correctly when writing
-    /// the diagram lines in reverse order.
-    ///
-    /// Also see [`Config::inverted_annotations`](crate::Config::inverted_annotations) to modify the layout algorithm to
-    /// draw annotations correctly if you are also rendering annotations.
-    ///
-    /// ## Example
-    /// Inverting has the following effect (the third column has reversed lines).
-    /// ```txt
-    /// 0         0          8
-    /// ├┬╮       ├┴╯       7╰╮
-    /// │1├╮      │1├╯      │6╰╮
-    /// ││2│      ││2│      ││5╰╮
-    /// │3│├╮     │3│├╯     │││4│
-    /// │╭╯││  →  │╰╮││  ↺  ││╰┤│
-    /// ││╭┤│     ││╰┤│     │╰╮││
-    /// │││4│  →  │││4│  ↺  │3│├╯
-    /// ││5╭╯     ││5╰╮     ││2│
-    /// │6╭╯      │6╰╮      │1├╯
-    /// 7╭╯       7╰╮       ├┴╯
-    ///  8         8        0
-    /// ```
-    pub const fn invert(mut self) -> Self {
-        self.charset = self.charset.invert();
-        self
-    }
-
-    /// The margin at the beginning of each annotation.
-    ///
-    /// This is the number of characters written at the beginning of the annotation line to create
-    /// a gap between the annotation and the branch diagram lines.
-    ///
-    /// The default value is `1`.
-    ///
-    /// ## Example
-    /// Here is an example going from annotation margin `0` to `2`.
-    /// ```txt
-    /// 0             0
-    /// ├┬╮           ├┬╮
-    /// │1├╮L0    →   │1├╮  L0
-    /// ││││L1        ││││  L1
-    /// ││2│          ││2│
-    /// │3│├╮L0   →   │3│├╮  L0
-    /// │╭╯││         │╭╯││
-    /// ││╭┤│         ││╭┤│
-    /// │││4│L0   →   │││4│  L0
-    /// │││╭╯L1       │││╭╯  L1
-    /// ││││ L2       ││││   L2
-    /// ││5│          ││5│
-    /// │6╭╯          │6╭╯
-    /// 7╭╯           7╭╯
-    ///  8             8
-    /// ```
-    pub const fn annotation_margin(mut self, annotation_margin: usize) -> Self {
-        self.annotation_margin = annotation_margin;
-        self
-    }
-
-    /// The minimum left justification of the annotations.
-    ///
-    /// Annotation lines will never begin earlier than this value, but may begin later than this
-    /// value if the row width plus annotation margin exceeds the justification. Note that the
-    /// justficiation is the number of characters, rather than columns, and therefore may need to be
-    /// adjusted if the gutter width is not zero.
-    ///
-    /// The default value is `0`.
-    ///
-    /// ## Example
-    /// Here is an example going from justification 0 to 5, with an annotation margin of 1.
-    /// Observe that:
-    ///
-    /// - `L0` is not shifted, since the justification refers to the start of the line after the
-    ///   margin
-    /// - `L1` still starts 1 character to the right of the justification since the diagram is too wide.
-    /// - `L2` is shifted right by two characters, so that it begins 5 characters from the
-    ///   boundary.
-    ///
-    /// right by one character, whereas the other annotations are not moved.
-    /// ```txt
-    /// 0             0
-    /// ├┬╮           ├┬╮
-    /// │1├╮ L0   →   │1├╮ L0
-    /// ││2│          ││2│
-    /// │3│├╮         │3│├╮
-    /// │╭╯││         │╭╯││
-    /// ││╭┤│         ││╭┤│
-    /// │││4│ L1  →   │││4│ L1
-    /// ││5╭╯         ││5╭╯
-    /// │6╭╯          │6╭╯
-    /// 7╭╯           7╭╯
-    ///  8 L2     →    8   L2
-    ///               ---->
-    /// ```
-    pub const fn annotation_justification(mut self, annotation_justification: usize) -> Self {
-        self.annotation_justification = annotation_justification;
-        self
-    }
-
-    /// The gap between columns in the branch diagram.
-    ///
-    /// The default value is `0`.
-    ///
-    /// ## Example
-    /// Setting the gutter width to `1` has the following effect.
-    /// ```txt
-    /// 0         0     
-    /// ├┬╮       ├─┬─╮  
-    /// │1├╮      │ 1 ├─╮
-    /// ││2│      │ │ 2 │
-    /// │3│├╮     │ 3 │ ├─╮
-    /// │╭╯││  →  │ ╭─╯ │ │
-    /// ││╭┤│     │ │ ╭─┤ │
-    /// │││4│  →  │ │ │ 4 │
-    /// ││5╭╯     │ │ 5 ╭─╯
-    /// │6╭╯      │ 6 ╭─╯
-    /// 7╭╯       7 ╭─╯   
-    ///  8          8
-    ///            ^ ^ ^ ^ gutters
-    /// ```
-    pub const fn gutter_width(mut self, gutter_width: usize) -> Self {
-        self.gutter_width = gutter_width;
-        self
-    }
-
-    /// Whether merge lines should go above diagram lines.
-    ///
-    /// Setting this value to `true` typically requires a
-    /// [gutter width](Self::gutter_width) of at least 1 for diagram clarity.
-    ///
-    /// The default value is `false`.
-    ///
-    /// ## Example
-    /// With a gutter width of 1, setting this to `true` has the following effect:
-    /// ```txt
-    /// 0             0
-    /// ├─╮           ├─╮
-    /// │ 1           │ 1
-    /// │ ├─╮         │ ├─╮
-    /// │ 2 ╰───╮     │ 2 ╰───╮
-    /// │ ├─┬─╮ │     │ ├─┬─╮ │
-    /// ├─│─╯ │ │  →  ├───╯ │ │
-    /// 3 │ ╭─╯ │     3 │ ╭─╯ │
-    /// │ │ 4 ╭─╯     │ │ 4 ╭─╯
-    /// │ ╰─╮ │       │ ╰─╮ │
-    /// ├─╮ │ │       ├─╮ │ │
-    /// ├─│─╯ │    →  ├───╯ │
-    /// 5 │ ╭─╯       5 │ ╭─╯
-    /// ╭─┴─╯         ╭─┴─╯
-    /// 6             6
-    /// ```
-    pub const fn merge_over(mut self, merge_over: bool) -> Self {
-        self.merge_over = merge_over;
-        self
-    }
-}
-
-/// A set of characters to use in the diagram.
-#[derive(Debug, Clone, PartialEq, Eq)]
-pub struct Charset {
-    /// The ` ` character.
-    pub space: char,
-    /// The `╮` character.
-    pub down_and_left: char,
-    /// The `╭` character.
-    pub down_and_right: char,
-    /// The `╯` character.
-    pub up_and_left: char,
-    /// The `╰` character.
-    pub up_and_right: char,
-    /// The `─` character.
-    pub horizontal: char,
-    /// The `│` character.
-    pub vertical: char,
-    /// The `┬` character.
-    pub down_and_horizontal: char,
-    /// The `┴` character.
-    pub up_and_horizontal: char,
-    /// The `┤` character.╮
-    pub vertical_and_left: char,
-    /// The `├` character.
-    pub vertical_and_right: char,
-    /// The `┼` character.
-    pub vertical_and_horizontal: char,
-}
-
-impl Charset {
-    /// Using this character set in a diagram style.
-    pub const fn style(self) -> Style {
-        Style::new(self)
-    }
-
-    /// The rounded corners character set.
-    /// ```txt
-    /// ╯ ┴ ╰
-    /// ┤ ┼ ├ ─
-    /// ╮ ┬ ╭ │
-    /// ```
-    pub const fn rounded_corners() -> Self {
-        Self {
-            space: ' ',
-            down_and_left: '╮',
-            down_and_right: '╭',
-            up_and_left: '╯',
-            up_and_right: '╰',
-            horizontal: '─',
-            vertical: '│',
-            down_and_horizontal: '┬',
-            up_and_horizontal: '┴',
-            vertical_and_left: '┤',
-            vertical_and_right: '├',
-            vertical_and_horizontal: '┼',
-        }
-    }
-
-    /// The sharp corners character set.
-    /// ```txt
-    /// ┘ ┴ └
-    /// ┤ ┼ ├ ─
-    /// ┐ ┬ ┌ │
-    /// ```
-    pub const fn sharp_corners() -> Self {
-        Self {
-            down_and_left: '┐',
-            down_and_right: '┌',
-            up_and_left: '┘',
-            up_and_right: '└',
-            ..Self::rounded_corners()
-        }
-    }
-
-    /// The doubled lines character set.
-    /// ```txt
-    /// ╝ ╩ ╚
-    /// ╣ ╬ ╠ ═
-    /// ╗ ╦ ╔ ║
-    /// ```
-    pub const fn doubled_lines() -> Self {
-        Self {
-            space: ' ',
-            down_and_left: '╗',
-            down_and_right: '╔',
-            up_and_left: '╝',
-            up_and_right: '╚',
-            horizontal: '═',
-            vertical: '║',
-            down_and_horizontal: '╦',
-            up_and_horizontal: '╩',
-            vertical_and_left: '╣',
-            vertical_and_right: '╠',
-            vertical_and_horizontal: '╬',
-        }
-    }
-
-    /// Flip the orientation of the charset vertically.
-    ///
-    /// This swaps the `down_*` characters with the `up_*` characters. This is useful when you
-    /// want to render the branch diagram with the root at the bottom since
-    /// it will make the branch diagram lines connect correctly when the lines are written in
-    /// reverse order.
-    ///
-    /// Also see [`Config::inverted_annotations`](crate::Config::inverted_annotations) to modify the layout algorithm to
-    /// draw annotations correctly.
-    pub const fn invert(self) -> Self {
-        Self {
-            down_and_left: self.up_and_left,
-            down_and_right: self.up_and_right,
-            down_and_horizontal: self.up_and_horizontal,
-            up_and_left: self.down_and_left,
-            up_and_right: self.down_and_right,
-            up_and_horizontal: self.down_and_horizontal,
-            ..self
-        }
     }
 }
